requirements.txt
flask 
open
flask_cors
pandas
numpy
tiktoken
jupyter-tabnine